fateanother_rl/requirements.txt
torch>=2.0
numpy>=1.24
pyyaml>=6.0
tensorboard>=2.14
